gdb: Fix type printing for C functions
Currently, when trying to print the types of a C function, GDB doesn't
strip the typedefs away. This is a problem because dwarf2_physname
expects to get the canonical name for functions. The reason this isn't
very apparent is because dwarf2_physname usually defers to demangling
the linkage_name when available, however a few tests have been written
to comply with this bug, that are changed in this commit.