gdb/cpp: delay demangled name calculation when linkage_name is unavailable
When GDB doesn't have the linkage name of a symbol, it attempts to
calculate one based on the available parameters, however, some
parameters might have typedefs and the current implementation does not
strip typedefs.

Unfortunately we can't just strip the typedefs as we read the symbols,
because some of the types may not have been fully read. One such example
is using gcc 12.2.1 and compiling the following function:

void foo (my_int *i)

since the pointer type description comes after function description, and
attempting to strip the typedef will crash GDB. Instead, we add these
types to the delayed_physname list already present in the cu struct.

We keep the original call to dwarf2_physname because we need some form
of name to some symbols while finishing the new_symbol calculation, so
we have to keep the original call and overwrite the linkage name
afterwards.