WIP fix for the name canonicalization problem
when testing gdb.linespec/cp-completion-aliases.exp with clang we see
failures. The problem is, that is the correct output, and GDB is instead
failing to properly canonicalize the name of functions. This patch fixes
it, but it causes some 2 thousand regressions... It does this by
changing how typedefs are replaced, using "check_typedef" inside
"replace_typedefs", but it needed a small change to struct type to make
all pointer types have some basic name().

It also doesnt properly calculate that basic name, instead using the C
convention and not working for pointers to functions